+++ test/inputs.test.ts
@@ -1,3 +1,5 @@
+import * as yaml from "js-yaml";
+
 import * as core from "./mocks/@actions/core.mock";
 
 jest.mock("@actions/core", () => core);
@@ -44,9 +46,33 @@ describe("ActionConfig::constructor", () => {
   });
 
   test("construct with an object specifying dry-run", () => {
-    const config: RawActionConfig = { "dry-run": "true" };
+    const config: RawActionConfig = { "dry-run": true };
     expect(() => new ActionConfig(config)).not.toThrow();
   });
+
+});
+
+describe("ActionConfig.getSvgoOptionsPath", () => {
+
+  const testPaths = test.each([".svgo.yml", "foo.yml", "in/folder/config.yml"]);
+
+  testPaths("svgo-options is set (to '%s') in the workflow file", (path) => {
+    const instance: ActionConfig = new ActionConfig();
+    core.getInput.mockReturnValueOnce(path);
+
+    const result = instance.getSvgoOptionsPath();
+    expect(result).toBe(path);
+  });
+
+  testPaths("svgo-options is set (to '%s') in the config object", (path) => {
+    const instance: ActionConfig = new ActionConfig({ "svgo-options": path });
+    core.getInput.mockReturnValueOnce(`dir/${path}`);
+
+    const result = instance.getSvgoOptionsPath();
+    expect(result).toBe(path);
+  });
+
+  afterAll(core.getInput.mockReset);
 
 });
 
@@ -84,7 +110,8 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   test("dry-run is `false` in the config object", () => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": "false" });
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: false");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("true");
 
     const result = instance.isDryRun();
@@ -92,7 +119,26 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   test("dry-run is `true` in the config object", () => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": "true" });
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: true");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
+    core.getInput.mockReturnValueOnce("false");
+
+    const result = instance.isDryRun();
+    expect(result).toBe(true);
+  });
+
+  test("dry-run is `'false'` in the config object", () => {
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'false'");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
+    core.getInput.mockReturnValueOnce("true");
+
+    const result = instance.isDryRun();
+    expect(result).toBe(false);
+  });
+
+  test("dry-run is `'true'` in the config object", () => {
+    const rawConfig: RawActionConfig = yaml.safeLoad("dry-run: 'true'");
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("false");
 
     const result = instance.isDryRun();
@@ -100,12 +146,15 @@ describe("ActionConfig.isDryRun", () => {
   });
 
   testNonBoolean("dry run is `'%s'` in the config object", async (value) => {
-    const instance: ActionConfig = new ActionConfig({ "dry-run": value });
+    const rawConfig: RawActionConfig = yaml.safeLoad(`dry-run: '${value}'`);
+    const instance: ActionConfig = new ActionConfig(rawConfig);
     core.getInput.mockReturnValueOnce("false");
 
     const result = instance.isDryRun();
     expect(result).toBe(true);
     expect(core.info).toHaveBeenCalledTimes(1);
   });
+
+  afterAll(core.getInput.mockReset);
 
 });